#Dreamsthatwalk
Dreaming into existence every day, living my life for God and praying for the skies to stay clear.  
Let's join us all to end the war on drugs,politics and other things that make us angry about being on earth. Let's make it what we want to think  is all that we need.